Create bookprinting
What is UV inkjet printing ink in book printing?

UV ink has a unique curing method, that is, it does not rely on permeation and evaporation to cure quickly, and it has been widely used in inkjet printing. UV inks have three major characteristics: energy-saving drying, wide printing adaptability to substrates (can be printed on almost all materials), and fast curing can shorten the follow-up process time. But UV ink is expensive and the cost is too high, which directly affects the printing price of UV inkjet printing. At this moment, you can contact us freely to learn about how to make the cheap book printing China.
The flexibility of UV ink is not good. If it curls too much, it will break; and its price is higher than that of solvent-based ink. At present, in the field of inkjet printing, UV inkjet printing inks are mainly used for wide-format graphics and image printing with a small number of copies. Because the number of copies is small, there is less restriction on the ink and substrate, and the advantages of UV ink printing can be used. The real advantage of UV inks is that they can be dried instantaneously under the action of UV lamps in long-run printing.
Because of this, UV ink may become the first choice in high-speed inkjet printing, especially when inkjet technology enters the field of offset printing such as packaging printing. From the perspective of environmental protection and the adaptability of substrates, UV inks are likely to become the mainstream of the large-format color market in the future. The recent UV-LED technology is to cure the ink under the UV light emitted by the LED light source. This type of ink is called UV-LED curing ink.
Because the cost and maintenance of high-energy-consuming curing lamps such as UV light sources have always been the biggest obstacle to the performance of the printer, if the temperature of the curing light source is set too high, it will have a thermal effect on foam and other materials (the heat will be concentrated on the surface of the material). Like in self-adhesive printing, these materials are susceptible to the adverse effects of high energy consumption lamps, thus losing their due role. The UV-LED technology has low energy consumption and low heat dissipation, and can significantly reduce the heat generated during the curing process. Therefore, UV inkjet printing can be performed on materials such as stickers and thin plastics.
Now, UV-LED technology has begun to be applied to large-format inkjet printers. At present, most printers in China use UV curing inkjet printing systems for printing.
If you want to find the custom printing China, we are happy to provide you with various services and assistance.